+++ src/fx.js
@@ -208,12 +208,17 @@ SVG.extend(SVG.FX, {
   }
   // Add animatable attributes
 , attr: function(a, v, n) {
-    if (typeof a == 'object')
+    if (typeof a == 'object') {
       for (var key in a)
         this.attr(key, a[key])
     
-    else
-      this.attrs[a] = { from: this.target.attr(a), to: v }
+    } else {
+      var from = this.target.attr(a)
+
+      this.attrs[a] = SVG.Color.isColor(from) ?
+        new SVG.Color(from).morph(v) :
+        { from: from, to: v }
+    }
     
     return this
   }
@@ -410,31 +415,10 @@ SVG.extend(SVG.FX, {
         .plus(new SVG.Number(o.from)) :
     
     /* color recalculation */
-    o.to && (o.to.r || SVG.Color.test(o.to)) ?
-      this._color(o, pos) :
+    o instanceof SVG.Color ? o.at(pos) :
     
     /* for all other values wait until pos has reached 1 to return the final value */
     pos < 1 ? o.from : o.to
-  }
-  // Private: tween color
-, _color: function(o, pos) {
-    var from, to
-    
-    /* normalise pos */
-    pos = pos < 0 ? 0 : pos > 1 ? 1 : pos
-    
-    /* convert FROM */
-    from = new SVG.Color(o.from)
-    
-    /* convert TO hex to rgb */
-    to = new SVG.Color(o.to)
-    
-    /* tween color and return hex */
-    return new SVG.Color({
-      r: ~~(from.r + (to.r - from.r) * pos)
-    , g: ~~(from.g + (to.g - from.g) * pos)
-    , b: ~~(from.b + (to.b - from.b) * pos)
-    }).toHex()
   }
   
 })